requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
schedule>=1.1.0
streamlit>=1.20.0
lxml>=4.9.0
python-dotenv>=0.19.0
